delete record mod and move things around
