+++ anchor/network/src/handshake/mod.rs
@@ -1,4 +1,4 @@
 pub mod behaviour;
 mod codec;
-pub mod record;
 pub mod types;
+pub mod envelope;
