--- anchor/network/src/handshake/record/mod.rs
@@ -1,2 +1,0 @@
-pub mod envelope;
-pub mod signing;
